Separated Functions To New File
